added so in account can message about rental

+++ backend/src/main/java/waxwing/campusbike/types/dto/BikeUpdateRequest.java
@@ -78,7 +78,7 @@ public class BikeUpdateRequest {
     this.longitude = longitude;
   }
 
-  public double getid() {
+  public long getid() {
     return id;
   }
